pyproject.toml
[tool.poetry]
name = "back"
version = "1.0.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.9"
Django = "^3.0.2"
psycopg2-binary = "^2.8"
gunicorn = "^19.9"
django-environ = "^0.4.5"
dj-database-url = "^0.5.0"
django-solo = "^1.1"
django-redis = "^5.0.0"
redis = "^3.5.3"
django-celery-beat = "^2.2.1"
django-celery-results = "^2.2.0"
validators = "^0.18.2"
Pillow = "^9.0.0"
celery = "^5.2.6"
flower = "^1.0.0"
django-admin-sortable = "^2.3"
pyTelegramBotAPI = "^4.5.1"

[tool.poetry.dev-dependencies]
isort = "^5.7.0"
black = "^20.8b1"
parameterized = "^0.8.1"
pre-commit = "^2.10.1"
autoflake = "^1.4"
django-dynamic-fixture = "^3.1.1"
seed-isort-config = "^2.2.0"
bandit = "^1.7.0"
flake8 = "^3.9.2"
mock = "^4.0.3"
requests-mock = "^1.9.3"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'DJANGO', 'FIRSTPARTY', 'LOCALFOLDER']
known_django = ['django']
known_third_party = "PIL,adminsortable,celery,dj_database_url,environ,redis,solo,telebot,validators"

[tool.black]
line-length = 120
exclude = '''
/(
  \.git
  | tmp/
  | data/
)/
'''


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
